Refactor separate text_sensor status component into nested component
