+++ components/philips_series_2200/text_sensor/status_sensor.cpp
