+++ example.yaml
@@ -29,7 +29,7 @@ philips_series_2200:
   id: philip
 
 text_sensor:
-  - platform: philips_status_sensor
+  - platform: philips_series_2200
     controller_id: philip
     id: status
     name: "Status"
